pyproject.toml
[project]
name = "school-report-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.32.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pydantic>=2.13.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
